Use URL and prefix (directory) as git cache keys
Using URL instead of git revisions should ensure that the cache
doesn't grow indefinitely. With git revisions, the cache would be soon
full of mostly old revisions. The advantage using git revision is
that, for example, different branches of a repo can be cached
simultaneously. I hope, this is not the common usage pattern. And if
it is, one can always set XDG_CACHE_HOME variable to use different
caches for different branches.